% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emcov.R
\name{em.cov}
\alias{em.cov}
\title{EM algorithm using covariance matrix parameterization}
\usage{
em.cov(
  dat,
  max.iter = 500,
  tol = 1e-05,
  start = c("diag", "pairwise", "listwise", "full"),
  debug = 0,
  ...
)
}
\arguments{
\item{dat}{blah}

\item{max.iter}{blah}

\item{tol}{blah}

\item{start}{blah}

\item{debug}{blah}

\item{...}{blah}
}
\description{
EM algorithm using covariance matrix parameterization
}
\details{
Additional details...
}
\examples{
\dontrun{
  library(psych)
  data(bfi)
  test <- em.cov(bfi[,1:25])
}

}
