% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll.R
\name{ICggm}
\alias{ICggm}
\title{Information criteria for GGM under missing data}
\usage{
ICggm(
  dat,
  index = c("EBIC", "BIC", "AIC", "nll"),
  mu,
  K = NULL,
  sig = NULL,
  N = NULL,
  gam = 0.5,
  tol = 1e-32
)
}
\arguments{
\item{dat}{Data frame or matrix that contains the raw data.}

\item{index}{Which information criteria to compute? See details.}

\item{mu}{Numeric vector of means.}

\item{K}{Precision matrix. If not provided, \code{sig} is required.}

\item{sig}{Covariance matrix. If not provided, \code{K} is required.}

\item{N}{Optional sample size for use in information criteria computations.}

\item{gam}{Single numeric value for \eqn{\gamma} in EBIC computations.}

\item{tol}{Tolerance for declaring edges to be zero. Anything in absolute value
below \code{tol} will be considered zero and will not count as a parameter.}
}
\value{
Single numeric value corresponding to the selected information criterion.
}
\description{
Information criteria for GGM under missing data
}
\details{
Computes information criterion (EBIC, BIC, AIC, or negative log-likelihood)
  for the given dataset, which may contain missing values, under the provided
  means (\code{mu}) and either precision matrix (\code{K}) or covariance matrix
  (\code{sig}) and assuming multivariate normality. Provide only \code{K} or
  \code{sig}, not both. Sample size computations are based on average pair-wise
  observations, unless a value for \code{N} is provided. Note that \code{gam}
  is the value for \eqn{\gamma} in EBIC computations. The typical choice is .5,
  but will be set to 0 if BIC is desired. Equations used are as follows:
  
  \deqn{EBIC = -2ll + E\log(N) + 4\gamma E \log(P)}
  \deqn{AIC = -2ll + 2E}
  
  Where \eqn{ll} is the log-likelihood, \eqn{E} is the number of edges (i.e., number
  of estimated parameters, though mean structure is ignored), \eqn{N} is the sample
  size, and \eqn{P} is the number of nodes (i.e., columns in the dataset).
}
