% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll.R
\name{nll.parpen}
\alias{nll.parpen}
\title{Penalized negative log-likelihood, parameter vector as input}
\usage{
nll.parpen(
  p,
  dat,
  param = c("prec", "cov"),
  fix.idx = NULL,
  fix.val = NULL,
  lambda = NULL,
  pendiag = FALSE
)
}
\arguments{
\item{p}{Parameter vector for means, then precision or covariance matrix
(rows of lower triangular portion).}

\item{dat}{Data frame or matrix that contains the raw data.}

\item{param}{Parameterization to use: Precision matrix or covariance matrix.}

\item{fix.idx}{Optional vector of parameter index numbers that will be fixed
to particular values.}

\item{fix.val}{Optional vector of parameter values; should match length of
\code{fix.idx}.}

\item{lambda}{Optional numeric value for tuning parameter for L1 penalty
term.}

\item{pendiag}{Boolean value indicating whether to penalize the diagonal of
the precision matrix.}
}
\value{
Single numeric value corresponding to the negative penalized
  log-likelihood.
}
\description{
Penalized negative log-likelihood, parameter vector as input
}
\details{
Assuming multivariate normality and accommodating missing data by
  using all available information (e.g., as would be used in direct maximum
  likelihood), computes value of L1 penalized log-likelihood.
  Off-diagonal elements of the matrix (precision or covariance)
  are penalized, with the option to also penalize the diagonal elements.
  
  Let \eqn{\theta} contain all model parameters, and \eqn{l()} be the
  log-likelihood, then the negative penalized log-likelihood is:
  \deqn{-l(\theta) + \lambda \sum |k_{jj'} |}
  
  where \eqn{\lambda} is the tuning parameter and the sum is over all unique
  elements in the matrix (precision or covariance), which, depending on the
  options chosen may or may not include the diagonal.
  
  This function uses a parameter vector as input, thereby possibly
  facilitating optimization. It also supports fixing particular parameters
  indexed by \code{fix.idx} to particular values in \code{fix.val}.
}
