% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nllcov}
\alias{nllcov}
\title{Negative log-likelihood, covariance matrix parameterization}
\usage{
nllcov(dat, mu, sig)
}
\arguments{
\item{dat}{(matrix) data.}

\item{mu}{(vector) means.}

\item{sig}{(matrix) covariance matrix.}
}
\value{
Numeric value of the negative log-likelihood.
}
\description{
Negative log-likelihood, covariance matrix parameterization
}
\details{
Computes the negative log-likelihood under missing data for the given value of means and covariances.
}
