% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nllprec}
\alias{nllprec}
\title{Negative log-likelihood, precision matrix parameterization}
\usage{
nllprec(dat, mu, K)
}
\arguments{
\item{dat}{(matrix) data.}

\item{mu}{(vector) means.}

\item{K}{(matrix) precision matrix.}
}
\value{
Numeric value of the negative log-likelihood.
}
\description{
Negative log-likelihood, precision matrix parameterization
}
\details{
Computes the negative log-likelihood under missing data for the given value of means and precision matrix.
}
