% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprec.R
\name{em.prec}
\alias{em.prec}
\title{EM algorithm originally in Stadler & Buhlmann (2012)}
\usage{
em.prec(
  dat,
  max.iter = 500,
  tol = 1e-05,
  start = c("diag", "pairwise", "listwise", "full"),
  debug = 0,
  ...
)
}
\arguments{
\item{dat}{blah}

\item{max.iter}{blah}

\item{tol}{blah}

\item{start}{blah}

\item{debug}{blah}

\item{...}{blah}
}
\description{
EM algorithm originally in Stadler & Buhlmann (2012)
}
\details{
Additional details...
}
\examples{
\dontrun{
  library(psych)
  data(bfi)
  test <- em.prec(bfi[,1:25])
}
}
